Update GLAD to version 4.4 to support double percision and change the variable types from float to double, so that we can zoom in further

+++ src/shaders/fragment.glsl
@@ -2,9 +2,9 @@
 in vec3 colorVertex;
 out vec4 fragmentColor;
 
-uniform float fractalScalar;
-uniform float zoomingOffsetX;
-uniform float zoomingOffsetY;
+uniform double fractalScalar;
+uniform double zoomingOffsetX;
+uniform double zoomingOffsetY;
 
 const int ITERATIONS = 500;
 
@@ -16,15 +16,15 @@ vec3 hsv2rgb(vec3 c)
     return c.z * mix(K.xxx, clamp(p - K.xxx, 0.0, 1.0), c.y);
 }
 
-int mandelbrot(int iter, vec2 pixel) {
+int mandelbrot(int iter, dvec2 pixel) {
     // convert pixel to range (0, 2)
-    vec2 c = vec2(pixel.x * fractalScalar*2 + zoomingOffsetX, pixel.y * fractalScalar*2 - zoomingOffsetY);
+    dvec2 c = dvec2(pixel.x * fractalScalar*2 + zoomingOffsetX, pixel.y * fractalScalar*2 - zoomingOffsetY);
 
     // I use vec2 to represent complex numbers, where x is the real part and y the imaginary part
-    vec2 z = vec2(0.0);
+    dvec2 z = dvec2(0.0);
     for (int i = 0; i < iter; i++) {
         // square z
-        float x = z.x;
+        double x = z.x;
         z.x = x * x - z.y * z.y;
         z.y = x * z.y * 2;
 
@@ -38,7 +38,7 @@ int mandelbrot(int iter, vec2 pixel) {
 }
 
 void main() {
-    vec2 uv = vec2(colorVertex.xy);
+    dvec2 uv = dvec2(colorVertex.xy); // omit the z axis and convert the vector into a double vector
     int exitedOnIteration = mandelbrot(ITERATIONS, uv);
     float clampedColor = exitedOnIteration/5 * 0.1;
     vec3 hsvColor = hsv2rgb(vec3(clampedColor, 0.5, (clampedColor >= 0.9 ? 0.9 : clampedColor)));
